pyproject.toml
[project]
name = "gradia-backend"
version = "0.1.0"
description = "gradia backend service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.115.12",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "selenium>=4.31.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.5.2",
    "httpx>=0.27.0",
    "requests>=2.32.3",
    "locust>=2.34.1",
    "psutil>=7.0.0",
    "datetime>=5.5",
    "webdriver-manager>=4.0.2",
    "playwright>=1.51.0",
    "google-auth>=2.40.1",
    "python-jose[cryptography]>=3.4.0",
    "firebase-admin>=6.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.1",
]
